fixed tests import sequence
