+++ stored_messages/tests/test_redis_backend.py
@@ -1,4 +1,4 @@
-from . import BaseTest
+from .base import BaseTest
 
 import unittest
 import mock
