+++ stored_messages/tests/test_restapi.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-from . import BackendBaseTest
+from .base import BackendBaseTest
 
 from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
